+++ tests/schema_test.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 import collections
+import io
 
 from pytest import fixture, mark, raises
 
@@ -465,6 +466,19 @@ def test_complete(fx_test_doc):
     assert is_partially_loaded(doc)
     complete(doc)
     assert not is_partially_loaded(doc)
+
+
+def test_complete_terminates_streaming():
+    f = io.BytesIO(b'''
+        <nstest xmlns="http://earthreader.github.io/"
+                xmlns:nst="https://github.com/earthreader/libearth">
+        </nstest>
+    ''')
+    doc = read(XmlnsDoc, f)
+    complete(doc)
+    f.close()
+    assert doc.samens_attr is None
+    assert doc.otherns_attr is None
 
 
 @fixture